ratsd-token-v2: move claims to separate file
